+++ src/creature.h
@@ -10,6 +10,7 @@
 #include "map.h"
 #include "position.h"
 #include "tile.h"
+#include "luacreature.h"
 
 class Condition;
 class Container;
@@ -343,7 +344,6 @@ public:
 		}
 	}
 
-protected:
 	virtual bool useCacheMap() const { return false; }
 
 	struct CountBlock_t
@@ -439,6 +439,7 @@ protected:
 	friend class Game;
 	friend class Map;
 	friend class LuaScriptInterface;
+	friend class luaCreature;
 };
 
 #endif // FS_CREATURE_H
